U3 participation updates
git-svn-id: http://source.usc.edu/svn/scec_vdo_vtk@292 25aa0b20-aa2b-420e-afc3-456cccd12aca

+++ trunk/commons/src/org/scec/vtk/commons/legend/LegendUtils.java
@@ -36,8 +36,8 @@ public class LegendUtils {
 	public static LegendItem buildColorBarLegend(Plugin source, CPT cpt, String title, double x, double y) {
 		Preconditions.checkNotNull(cpt, "CPT cannot be null");
 		vtkScalarBarActor scalarBar = new vtkScalarBarActor();
-		int minValue = (int)cpt.getMinValue();
-		int maxValue = (int)cpt.getMaxValue();
+		double minValue = cpt.getMinValue();
+		double maxValue = cpt.getMaxValue();
 		Color minColor = cpt.getMinColor();
 		Color maxColor = cpt.getMaxColor();
 		int numTicks = cpt.size()+1;
